Edited the bands with fit plots
This is easier to read, but now it looks like the bands are actually the fit functions so it looks like there is perfect agreement, which is not the case. I'll try to add grey bands which cover the parts of the range which is not included in the fit.