Added the smoothed e-vals inverse of the covariance matrix
This is an alternative to the svd decomposition of the covariance matrix which could be better but it doesn't help us since it is meant for datasets where the correlations are overestimated due to a small sample size. However our correlations are just large as is. It is from the paper by Michael and McKerrell from 1995